added an adapter class
fixed #3
