+++ src/ListenerAdapter.java
@@ -1,0 +1,64 @@
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
+
+public class ListenerAdapter implements KeyListener, ActionListener, MouseListener {
+
+	@Override
+	public void mouseClicked(MouseEvent e) {
+		
+
+	}
+
+	@Override
+	public void mousePressed(MouseEvent e) {
+		
+
+	}
+
+	@Override
+	public void mouseReleased(MouseEvent e) {
+		
+
+	}
+
+	@Override
+	public void mouseEntered(MouseEvent e) {
+		
+
+	}
+
+	@Override
+	public void mouseExited(MouseEvent e) {
+		
+
+	}
+
+	@Override
+	public void actionPerformed(ActionEvent e) {
+		
+
+	}
+
+	@Override
+	public void keyTyped(KeyEvent e) {
+		
+
+	}
+
+	@Override
+	public void keyPressed(KeyEvent e) {
+		
+
+	}
+
+	@Override
+	public void keyReleased(KeyEvent e) {
+		
+
+	}
+
+}
