+++ src/StickmanParkour.java
@@ -1,14 +1,9 @@
   
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.KeyEvent;
-import java.awt.event.KeyListener;
-
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 
 @SuppressWarnings("serial")
-public class StickmanParkour extends JPanel implements ActionListener, KeyListener {
+public class StickmanParkour extends JPanel {
 	
 	GamePanel panel;
 	JFrame frame;
@@ -31,26 +26,11 @@ public class StickmanParkour extends JPanel implements ActionListener, KeyListen
 		frame.add(panel);
 		frame.setVisible(true);
 		frame.setSize(width, height);
-		frame.addKeyListener(panel);
-		frame.addMouseListener(panel);
+		frame.addKeyListener(panel.getListener());
+		frame.addMouseListener(panel.getListener());
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		panel.startGame();
 
 	}
 
-	@Override
-	public void keyTyped(KeyEvent e) {
-	}
-
-	@Override
-	public void keyPressed(KeyEvent e) {
-	}
-
-	@Override
-	public void keyReleased(KeyEvent e) {
-	}
-
-	@Override
-	public void actionPerformed(ActionEvent e) {
-	}
 }
